Initial Commit of A simple code style sheet, and the chassis class.
The chassis class is where we will have all our motors and sensors. I felt like this would be nice to have so we all can use ONE config file and not half to change it depending on the OPMode. All sensors, motors, servos will be public in this class and will get registered when we call init from any OPMode.